Revert "Fix/skip suspended cards in graphs context retrievability (#3518)"
This reverts commit 939cc5a268108eb518ebfeff3c9a21b3bb57be8b.

+++ rslib/src/browser_table.rs
@@ -115,8 +115,8 @@ impl Card {
         } else if self.is_due_in_days() {
             Some(
                 TimestampSecs::now().adding_secs(
-                    (self.original_or_current_due() as i64 - timing.days_elapsed as i64)
-                        .saturating_mul(86400),
+                    ((self.original_or_current_due() - timing.days_elapsed as i32)
+                        .saturating_mul(86400)) as i64,
                 ),
             )
         } else {
@@ -134,9 +134,9 @@ impl Card {
             )
         } else {
             self.due_time(timing).map(|due| {
-                (due.adding_secs(-86_400 * self.interval as i64)
-                    .elapsed_secs()
-                    / 86_400) as u32
+                due.adding_secs(-86_400 * self.interval as i64)
+                    .elapsed_secs() as u32
+                    / 86_400
             })
         }
     }
